+++ src/locales/nl/PAGES/AUDIT.json
@@ -34,7 +34,7 @@
     "TITLE": "Stap 4: Toets de sample",
     "UNDERSTAND": "Toelichting",
     "UNTESTED": "{{critCount}} niet getest",
-    "VIEW_IN_REPORT": "Toon in rapport",
+    "VIEW_IN_REPORT": "In rapport",
     "LABEL_VERSION": "Versie van WCAG",
     "LABEL_LEVEL": "Conformanceniveau"
 }
